# apps/mobile_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NTUBTOB Mobile API",
    "version": "1.0.0",
    "description": "Canonical native contract. Basic remains isolated; bounded Officer/Admin reads are projected from the fresh active Person principal. Tokens, provider assertions, nonces, idempotency keys and secrets are redacted from logs and error responses."
  },
  "servers": [{"url": "/api/v1"}],
  "security": [{"bearerAuth": []}],
  "paths": {
    "/auth/line/exchange": {
      "post": {
        "security": [],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LineExchangeRequest"}}}},
        "responses": {"201": {"description": "Device session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionEnvelope"}}}}, "202": {"description": "Pending identity; review-only credential issued and no mobile session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PendingReviewEnvelope"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/IdentityUnavailable"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}, "429": {"$ref": "#/components/responses/RateLimited"}, "503": {"$ref": "#/components/responses/Unavailable"}}
      }
    },
    "/auth/google/exchange": {
      "post": {
        "security": [],
        "description": "Accepts only a raw Google ID token. The backend verifies signature, Google issuer, expiry and an explicitly configured audience allowlist, then uses only verified sub as provider_subject. Email, name and avatar are not identity assertions and are not accepted.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GoogleExchangeRequest"}}}},
        "responses": {"201": {"description": "Device session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionEnvelope"}}}}, "202": {"description": "Pending identity; review-only credential issued and no mobile session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PendingReviewEnvelope"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/IdentityUnavailable"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}, "429": {"$ref": "#/components/responses/RateLimited"}, "503": {"$ref": "#/components/responses/Unavailable"}}
      }
    },
    "/auth/apple/exchange": {
      "post": {
        "security": [],
        "description": "Accepts only a single-use Apple authorization code, Apple ID token and original raw nonce on iOS. The backend first verifies the ID token, reserves the code hash, exchanges the code exactly once, correlates the returned ID-token subject, and persists only an independently encrypted provider refresh credential. Unknown exchange outcomes are terminal and must not be retried. Email, private relay email, name, user and profile hints are rejected. Missing lifecycle configuration or schema disables only Apple authentication.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AppleExchangeRequest"}}}},
        "responses": {"201": {"description": "Device session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionEnvelope"}}}}, "202": {"description": "Pending identity; review-only credential issued and no mobile session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PendingReviewEnvelope"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/IdentityUnavailable"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}, "503": {"$ref": "#/components/responses/Unavailable"}}
      }
    },
    "/auth/apple/notifications": {
      "post": {
        "security": [],
        "description": "Public provider callback. Accepts exactly one bounded application/x-www-form-urlencoded payload field containing an Apple-signed JWS. Signature, issuer, exact notification audience, issued-at, Unix-seconds event time and JTI bounds are verified before an idempotent receipt. consent-revoked and account-deleted atomically disable only the Apple identity and revoke its app sessions and encrypted provider credential; email forwarding events are receipt-only because provider email is not stored.",
        "requestBody": {"required": true, "content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/AppleNotificationRequest"}}}},
        "responses": {"204": {"description": "Verified event applied or exact JTI replay acknowledged"}, "400": {"$ref": "#/components/responses/Validation"}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "503": {"$ref": "#/components/responses/Unavailable"}}
      }
    },
    "/auth/identities": {
      "get": {"description": "List only safe provider labels and linked_at for the current Person; sensitive provider data is never returned.", "responses": {"200": {"description": "Redacted linked login methods"}, "401": {"$ref": "#/components/responses/Unauthenticated"}}}
    },
    "/auth/identity-link/candidates/{provider}": {
      "post": {"security": [], "description": "Verify an explicitly selected LINE, Google or Apple credential and issue a purpose-separated candidate credential for at most 300 seconds. LINE and Apple require the original raw nonce; Apple verifies the token nonce as lowercase SHA-256(raw nonce). Only verified provider sub selects the candidate; profile hints are rejected. The credential alone has no access or review authority.", "parameters": [{"$ref": "#/components/parameters/IdentityProvider"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IdentityLinkCandidateRequest"}}}}, "responses": {"201": {"description": "Candidate proof created"}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}}}
    },
    "/auth/identity-link/proofs/{provider}": {
      "post": {"security": [], "description": "Freshly verify the other linked active LINE, Google or Apple provider and bind it to the candidate transaction. LINE and Apple require the original raw nonce; profile hints are rejected.", "parameters": [{"$ref": "#/components/parameters/IdentityProvider"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IdentityLinkProofRequest"}}}}, "responses": {"201": {"description": "Fresh proof created"}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}}}
    },
    "/auth/identity-link/confirm": {
      "post": {"security": [], "description": "Explicit atomic cross-provider confirmation. Recovery returns a session only on the first successful confirmation; a lost-response replay returns already_linked without a second session.", "responses": {"200": {"description": "Linked or idempotently already linked"}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}}}
    },
...
    "/auth/refresh": {
      "post": {
        "security": [],
        "parameters": [{"$ref": "#/components/parameters/RefreshAttemptId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefreshRequest"}}}},
        "responses": {"200": {"description": "Rotated session envelope; same attempt and request replays exactly for 300 seconds", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionEnvelope"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}, "503": {"$ref": "#/components/responses/Unavailable"}}
      }
    },
    "/auth/logout": {
      "post": {"responses": {"204": {"description": "Current device session revoked idempotently"}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "503": {"$ref": "#/components/responses/Unavailable"}}}
    },
    "/me": {
      "get": {"responses": {"200": {"description": "Active Person and request-time server-computed capabilities", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Person"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/Forbidden"}, "503": {"$ref": "#/components/responses/Unavailable"}}},
      "patch": {"parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["display_name"], "properties": {"display_name": {"type": "string", "minLength": 1, "maxLength": 120}}}}}}, "responses": {"200": {"description": "Refreshed self Person with truthful changed and idempotent_replay", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileMutation"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}}}
    },
    "/auth/line/review": {
      "get": {"description": "Read only the pending identity bound to the review credential", "responses": {"200": {"description": "Own pending review", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PendingReview"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}}}
    },
    "/auth/line/review/messages": {
      "post": {"parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["body"], "properties": {"body": {"type": "string", "minLength": 1, "maxLength": 1000}}}}}}, "responses": {"200": {"description": "Applicant message appended within server rate limit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PendingReview"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/Validation"}}}
    },
...
    "/games/{game_id}": {
      "get": {"parameters": [{"$ref": "#/components/parameters/GameId"}], "responses": {"200": {"description": "Scoped game detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Game"}}}}, "401": {"$ref": "#/components/responses/Unauthenticated"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "503": {"$ref": "#/components/responses/Unavailable"}}}
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "short-lived app access token", "description": "Authorization header only; never cookie/query. Access TTL 900 seconds and maximum token length 4096 bytes."}},
...
    "schemas": {
...
      "LineExchangeRequest": {"type": "object", "additionalProperties": false, "required": ["id_token", "nonce", "login_attempt_id", "installation_id", "platform"], "properties": {"id_token": {"type": "string", "minLength": 1, "maxLength": 2048, "writeOnly": true}, "nonce": {"type": "string", "minLength": 16, "maxLength": 200, "writeOnly": true}, "login_attempt_id": {"type": "string", "minLength": 16, "maxLength": 200}, "installation_id": {"type": "string", "minLength": 16, "maxLength": 200}, "platform": {"type": "string", "enum": ["ios", "android"]}}},
...
      "SessionEnvelope": {"type": "object", "additionalProperties": false, "required": ["access_token", "refresh_token", "session_id", "expires_in"], "properties": {"access_token": {"type": "string", "writeOnly": true}, "refresh_token": {"type": "string", "writeOnly": true}, "session_id": {"type": "string", "maxLength": 64}, "expires_in": {"type": "integer", "const": 900}}},
...
    },
...
  }
}